Fix bug in decoding
It might be more efficient to use a larger buffer, but the underlying
gzip decoder should be buffered anyway. The previous implementation
wasn't checking if `length` was a multiple of 4, so there could be
some discarded bytes that weren't being read.

This version is a lot simpler, and maybe I can revisit if it's a bottleneck.